Rename ElectionUtil to Election
